feat: Add Routes page, route delete/update endpoints, inbox scheduled filter
- New /routes page with table view of all planned routes

- Route delete and update endpoints (sazinka.route.delete, sazinka.route.update)

- Inline crew assignment on Routes page

- Delete route button in Planner

- New S terminem (scheduled) segment filter in Inbox

- Fix migration 004 checksum (restore original REAL type)

+++ worker/migrations/004_add_crew_buffer.sql
@@ -1,3 +1,3 @@
 -- Add arrival buffer percent to crews table
 -- Buffer = percentage of preceding segment duration to arrive early before time window start
-ALTER TABLE crews ADD COLUMN arrival_buffer_percent DOUBLE PRECISION NOT NULL DEFAULT 10.0;
+ALTER TABLE crews ADD COLUMN arrival_buffer_percent REAL NOT NULL DEFAULT 10.0;
